fix: Phase 2 Task 4 - Fix Mermaid and JSON rendering in chat
🎯 PHASE 2 TASK 4 COMPLETE: Mermaid & JSON Rendering

## ✅ Improvements

### 1. Created Shared CodeElementBuilder
- Extracted _CodeElementBuilder from MarkdownPreviewWidget
- New file: src/client/lib/shared/presentation/widgets/markdown_builders/code_element_builder.dart
- Implements Strategy Pattern for code block routing:
  - Mermaid diagrams → MermaidView (rendered via mermaid.ink)
  - All other languages → HighlightView (syntax highlighting)
- Supports 185+ languages via flutter_highlighter
- Comprehensive DartDoc with usage examples

### 2. Updated SmartMessageRenderer
- Added md.ExtensionSet.gitHubFlavored for proper Markdown parsing
- Replaced basic MermaidBuilder with hybrid CodeElementBuilder
- Implements JSON auto-wrapping logic:
  - Detects pure JSON (starts with { or [)
  - Wraps in ```json code block if no backticks present
  - Enables syntax highlighting for raw JSON responses

### 3. Updated MarkdownPreviewWidget
- Removed duplicate _CodeElementBuilder class
- Now uses shared CodeElementBuilder from markdown_builders/
- Maintains consistent rendering between chat and document preview
- Removed unused mermaid_view import

## 🧪 Test Results

- ✅ Flutter analyze: No issues found
- ✅ All 929 tests passing (2 skipped)
- ✅ Test execution time: 43 seconds
- ✅ Coverage maintained at expected levels

## 📝 Technical Details

**Files Modified:**
- src/client/lib/features/chat/presentation/widgets/smart_message_renderer.dart
- src/client/lib/features/project_shell/presentation/widgets/markdown_preview_widget.dart

**Files Created:**
- src/client/lib/shared/presentation/widgets/markdown_builders/code_element_builder.dart

**Benefits:**
- Consistent Mermaid rendering across app (chat & preview)
- Syntax highlighting for JSON, YAML, Dart, Python, etc.
- DRY principle: Single source of truth for code rendering
- Better UX: Automatic formatting for raw JSON responses
