+++ src/client/lib/features/project_shell/presentation/widgets/markdown_preview_widget.dart
@@ -8,7 +8,7 @@ import 'package:flutter_markdown_plus/flutter_markdown_plus.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:markdown/markdown.dart' as md;
 
-import '../../../../shared/presentation/widgets/mermaid_view.dart'; // 👈 IMPORT DEL VISOR MERMAID
+import '../../../../shared/presentation/widgets/markdown_builders/code_element_builder.dart';
 import '../../../filesystem/presentation/notifiers/file_system_notifier.dart';
 import '../../../filesystem/presentation/providers/filesystem_providers.dart';
 import '../../infrastructure/services/project_progress_service.dart';
@@ -348,8 +348,8 @@ class _MarkdownPreviewWidgetState extends ConsumerState<MarkdownPreviewWidget> {
       padding: const EdgeInsets.all(24),
       extensionSet: md.ExtensionSet.gitHubFlavored,
       builders: {
-        'code': _CodeElementBuilder(),
-      }, // 👈 MANTENEMOS TU BUILDER HÍBRIDO
+        'code': CodeElementBuilder(),
+      },
       styleSheet: MarkdownStyleSheet.fromTheme(safeTheme).copyWith(
         p: const TextStyle(color: Color(0xFFC9D1D9), fontSize: 14, height: 1.6),
         h1: const TextStyle(
@@ -405,47 +405,4 @@ class _ToolbarButton extends StatelessWidget {
     splashRadius: 20,
     hoverColor: Colors.white.withValues(alpha: 0.1),
   );
-}
-
-/// Markdown code-block builder that routes Mermaid diagrams to [MermaidView]
-/// and all other languages to [HighlightView] for syntax colouring.
-class _CodeElementBuilder extends MarkdownElementBuilder {
-  @override
-  Widget? visitElementAfter(md.Element element, TextStyle? preferredStyle) {
-    var language = '';
-    if (element.attributes['class'] != null) {
-      final lg = element.attributes['class'] as String;
-      language = lg.startsWith('language-') ? lg.substring(9) : lg;
-    }
-    final codeContent = element.textContent.endsWith('\n')
-        ? element.textContent.substring(0, element.textContent.length - 1)
-        : element.textContent;
-
-    // Mermaid diagrams are rendered via mermaid.ink.
-    if (language == 'mermaid') {
-      return MermaidView(code: codeContent);
-    }
-
-    // All other languages use flutter_highlighter for syntax colouring.
-    return Container(
-      width: double.infinity,
-      margin: const EdgeInsets.symmetric(vertical: 4),
-      decoration: BoxDecoration(
-        color: const Color(0xFF161B22),
-        borderRadius: BorderRadius.circular(6),
-        border: Border.all(color: const Color(0xFF30363D)),
-      ),
-      child: HighlightView(
-        codeContent,
-        language: language,
-        theme: atomOneDarkTheme,
-        padding: const EdgeInsets.all(12),
-        textStyle: const TextStyle(
-          fontFamily: 'JetBrains Mono',
-          fontSize: 13,
-          height: 1.4,
-        ),
-      ),
-    );
-  }
 }
